api/services/__init__.py
from dotenv import load_dotenv

load_dotenv()

import services.analyse as analyse
import services.fakeNewsDetector as fakeNewsDetector
import services.googleUtils as googleUtils
import services.gpt as gpt
import services.webScrapper as webScrapper

__all__ = [
  "analyse", 
  "fakeNewsDetector",
  "googleUtils",
  "gpt",
  "webScrapper"
]
