api/requirements.txt
scikit-learn
flask
flask-cors
nltk
pandas
newspaper3k
openai
newspaper3k
google-api-python-client